+++ packages/core/test/integration/core/happyPath.test.ts
@@ -1480,92 +1480,6 @@ describe('Happy Path', () => {
     expect((await market.global()).exposure).to.equals(0)
   })
 
-  it('resolve market shortfall from insurance fund', async () => {
-    const POSITION = parse6decimal('10')
-    const POSITION_B = parse6decimal('1')
-    const COLLATERAL = parse6decimal('1000')
-    const { owner, user, userB, dsu, chainlink, marketFactory, insuranceFund } = instanceVars
-
-    const market = await createMarket(instanceVars)
-    await dsu.connect(user).approve(market.address, COLLATERAL.mul(1e12))
-    await dsu.connect(userB).approve(market.address, COLLATERAL.mul(1e12))
-
-    // Enable insurance fund as operator for market factory owner
-    await marketFactory.connect(owner).updateOperator(insuranceFund.address, true)
-
-    await market
-      .connect(user)
-      ['update(address,uint256,uint256,uint256,int256,bool)'](user.address, POSITION, 0, 0, COLLATERAL, false)
-    await market
-      .connect(userB)
-      ['update(address,uint256,uint256,uint256,int256,bool)'](userB.address, 0, POSITION_B, 0, COLLATERAL, false)
-
-    await chainlink.nextWithPriceModification(price => price.mul(10))
-
-    await settle(market, user)
-
-    const riskParameter = {
-      margin: parse6decimal('0.3'),
-      maintenance: parse6decimal('0.3'),
-      takerFee: {
-        linearFee: parse6decimal('0.001'),
-        proportionalFee: parse6decimal('0.0006'),
-        adiabaticFee: parse6decimal('0.0004'),
-        scale: parse6decimal('10000'),
-      },
-      makerFee: {
-        linearFee: parse6decimal('0.0005'),
-        proportionalFee: parse6decimal('0.0002'),
-        adiabaticFee: 0,
-        scale: parse6decimal('10000'),
-      },
-      makerLimit: parse6decimal('100000'),
-      efficiencyLimit: parse6decimal('0.2'),
-      liquidationFee: parse6decimal('10.00'),
-      utilizationCurve: {
-        minRate: 0,
-        maxRate: parse6decimal('5.00'),
-        targetRate: parse6decimal('0.80'),
-        targetUtilization: parse6decimal('0.80'),
-      },
-      pController: {
-        k: parse6decimal('40000'),
-        min: parse6decimal('-1.20'),
-        max: parse6decimal('1.20'),
-      },
-      minMargin: parse6decimal('500'),
-      minMaintenance: parse6decimal('500'),
-      staleAfter: 100000, // enable long delays for testing
-      makerReceiveOnly: false,
-    }
-    const parameter = {
-      fundingFee: parse6decimal('0.1'),
-      interestFee: parse6decimal('0.1'),
-      riskFee: 0,
-      maxPendingGlobal: 8,
-      maxPendingLocal: 8,
-      makerFee: parse6decimal('0.2'),
-      takerFee: parse6decimal('0.1'),
-      maxPriceDeviation: parse6decimal('0.1'),
-      closed: false,
-      settle: false,
-    }
-
-    await market.updateParameter(parameter)
-    await market.updateRiskParameter(riskParameter)
-
-    // ensure exposure is negative
-    expect((await market.global()).exposure).to.lt(0)
-
-    const insuranceFundSigner = await impersonate(insuranceFund.address)
-    await fundWallet(dsu, insuranceFundSigner)
-
-    // resolve market shortfall
-    await insuranceFund.connect(owner).resolve(market.address)
-
-    expect((await market.global()).exposure).to.equals(0)
-  })
-
   it('opens intent order w/ signer', async () => {
     const { owner, user, userB, userC, marketFactory, verifier, dsu } = instanceVars
